+++ resources/connector/connector.py
@@ -177,8 +177,8 @@ class ConnectorInputOutput(InputOutput):
       prompt = self.connector.file_watcher.process_changes()
       if prompt:
         changed_files = ", ".join(sorted(self.connector.file_watcher.changed_files))
-        wait_for_async(self.connector, self.connector.send_log_message("info", f"Detected changes in files: {changed_files}."))
-        wait_for_async(self.connector, self.connector.send_log_message("loading", "Processing changes..."))
+        wait_for_async(self.connector, self.connector.send_log_message("info", f"Detected AI request in files: {changed_files}."))
+        wait_for_async(self.connector, self.connector.send_log_message("loading", "Processing request..."))
         self.connector.loop.create_task(process_changes())
 
 def create_coder(connector):
@@ -230,7 +230,13 @@ class Connector:
     self.interrupted = False
 
     if watch_files:
-      self.file_watcher = FileWatcher(self.coder)
+      ignores = []
+      if self.coder.root:
+        ignores.append(self.coder.root + "/.gitignore")
+      if self.coder.repo.aider_ignore_file:
+        ignores.append(self.coder.repo.aider_ignore_file)
+
+      self.file_watcher = FileWatcher(self.coder, gitignores=ignores)
       self.file_watcher.start()
 
     try: